refactor: consolidate all hook handlers under `midtown hook`
Move task and ask hook handlers from `midtown coworker task-hook` and
`midtown coworker ask-hook` into `midtown hook task` and `midtown hook ask`.
Also remove the unused `midtown task hook create/update` subcommands.

All hook handlers now live in hooks.rs, and all run in the early-exit
path (before daemon connection) which is more correct since these hooks
should work even if the daemon is temporarily unavailable.
